Add files via upload
these are just some basic codes for an app i am trying to work on. i am still very green on the entire web app development but i am trying my best. this app is a hostel locator for students who are joining university and find it difficult to move around looking for hostels. this app allows them to browse and see different hostels around the universities they are studying at.